个人资料修改

+++ LADZhangWo/chaoshi/ChaoshiDetailViewController.m
@@ -27,7 +27,7 @@
     _contentWebView.delegate = self;
     [self.view addSubview:_contentWebView];
     
-    NSString *urlString = [SITEAPI stringByAppendingFormat:@"&mod=chaoshi&ac=showdetail&id=%d",self.goodsid];
+    NSString *urlString = [SITEAPI stringByAppendingFormat:@"&mod=chaoshi&ac=showdetail&id=%ld",(long)self.goodsid];
     [_contentWebView loadRequest:[NSURLRequest requestWithURL:[NSURL URLWithString:urlString]]];
 }
