+++ LADZhangWo/home/RecommendSliderView2.m
@@ -27,7 +27,7 @@
 }
 
 - (void)loadData{
-    [_afmanager GET:[SITEAPI stringByAppendingFormat:@"&mod=homepage&ac=showlist&groupid=%d&num=%d",_groupid,_dataCount] parameters:nil success:^(AFHTTPRequestOperation * _Nonnull operation, id  _Nonnull responseObject) {
+    [_afmanager GET:[SITEAPI stringByAppendingFormat:@"&mod=homepage&ac=showlist&groupid=%ld&num=%ld",(long)_groupid,(long)_dataCount] parameters:nil success:^(AFHTTPRequestOperation * _Nonnull operation, id  _Nonnull responseObject) {
         id array = [NSJSONSerialization JSONObjectWithData:(NSData *)responseObject options:NSJSONReadingAllowFragments error:nil];
         if ([array isKindOfClass:[NSArray class]]) {
             [self showImagesWithArray:array];
@@ -69,7 +69,7 @@
 - (void)imageClick:(UITapGestureRecognizer *)tap{
     NSInteger dataid = tap.view.tag;
     if ([self.tapDelegate respondsToSelector:@selector(showDetailWithID:andIdType:)]) {
-        NSString *idType = [_idTypes objectForKey:[NSString stringWithFormat:@"%d",dataid]];
+        NSString *idType = [_idTypes objectForKey:[NSString stringWithFormat:@"%ld",(long)dataid]];
         [self.tapDelegate showDetailWithID:dataid andIdType:idType];
     }
 }
